Add readonly flag to JSON configuration to read only config files

+++ faroe/json/src/main/java/org/n52/faroe/json/JsonConfiguration.java
@@ -61,6 +61,7 @@ public class JsonConfiguration implements Destroyable,
     private static final String WEB_INF_PATH = "WEB-INF";
     private String fileName = DEFAULT_FILE_NAME;
     private int writeTimeout = DEFAULT_WRITE_TIMEOUT;
+    private boolean readonly = false;
     private final ReadWriteLock lock = new ReentrantReadWriteLock();
     private final JsonNodeFactory nodeFactory = Json.nodeFactory();
     private ObjectNode configuration;
@@ -170,6 +171,15 @@ public class JsonConfiguration implements Destroyable,
     public void setWriteTimeout(int writeTimeout) {
         this.writeTimeout = writeTimeout;
     }
+    
+    /**
+     * Sets the flag to persist or not persist the settings, e.g. settings defined externally
+     *
+     * @param readonly the flag to persist settings or not
+     */
+    public void setReadonly(boolean readonly) {
+        this.readonly = readonly;
+    }
 
     /**
      * Sets the file name of this configuration.
@@ -229,16 +239,18 @@ public class JsonConfiguration implements Destroyable,
      * Actually persists the configuration.
      */
     private synchronized void persist() {
-        readLock().lock();
-        try {
-            LOG.debug("Writing configuration file");
-            try (FileOutputStream fos = new FileOutputStream(this.file)) {
-                Json.print(fos, this.configuration);
+        if (!readonly) {
+            readLock().lock();
+            try {
+                LOG.debug("Writing configuration file");
+                try (FileOutputStream fos = new FileOutputStream(this.file)) {
+                    Json.print(fos, this.configuration);
+                }
+            } catch (IOException e) {
+                throw new ConfigurationError("Could not persist configuration", e);
+            } finally {
+                readLock().unlock();
             }
-        } catch (IOException e) {
-            throw new ConfigurationError("Could not persist configuration", e);
-        } finally {
-            readLock().unlock();
         }
     }
